weather v0.0.1 Weather.CLI

Entry point of a program, main process is going here

Functions

main(argv)

Entry function

parse_args(argv)

Parse command line arguments passed to the application

Supported keys: —help, -h Prints usage message

Main usage: weather code - where code is a unique code of your location in US

Returns errors or help or code

process(errors)

Do som job based on passed parameter :help - diplay help message code - get data from weather.gov and pretty print it errors - output all errors