weather v0.0.1 Weather.PrettyPrinter

This module is used to prettyprint data parsed from xml

Functions

get_row_format(column_widths)

Gets a list of widths and returns a string that can be used to format a row

Example

iex> widths = [8, 9]
iex> Weather.PrettyPrinter.get_row_format widths
"~-8s | ~-9s~n"

get_width(values)

Takes a list of columns(each represented as a list of string values) and returns width of columns that should be enough to fit all contents.

Example

iex> col = ["kitt", "glad", "butterfly"]
iex> Weather.PrettyPrinter.get_width col
9

print(map)

Takes map with data to be printed and outputs a table

print_row(arg, format)

Takes a twodimensional tuple and prints its content using provided format

Example

iex> Weather.PrettyPrinter.print_row {:who_is_cool?, "me"}, "~-12s | ~-2s"
who_is_cool? | me