weather v0.0.1 Weather.XMLParser

This module parses raw string data into Elixir map

Functions

get_xpaths(list)

parse_xml(data)

Parses incoming string into map, if input string is incorrect - program crashes with error